ref : Email, Constellation, Connection DTO 클래스 요청 및 응답 패키지로 분리
Email, Constellation, Connection DTO 클래스에서 요청 및 응답 패키지로 추가 분할하였습니다.
