+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/constellation/dto/response/ArchiveDto.java
@@ -1,6 +1,6 @@
-package com.example.starlet_be.domains.constellation.dto;
+package com.example.starlet_be.domains.constellation.dto.response;
 
-import com.example.starlet_be.domains.connection.dto.ConnectionDto;
+import com.example.starlet_be.domains.connection.dto.response.ConnectionDto;
 import com.example.starlet_be.domains.star.dto.response.StarArchiveDto;
 import lombok.Builder;
 import lombok.Getter;
