+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/email/controller/EmailController.java
@@ -1,7 +1,7 @@
 package com.example.starlet_be.domains.email.controller;
 
 import com.example.starlet_be.domains.email.api.EmailAPI;
-import com.example.starlet_be.domains.email.dto.EmailAddressDto;
+import com.example.starlet_be.domains.email.dto.request.EmailAddressDto;
 import com.example.starlet_be.domains.email.service.EmailService;
 import com.example.starlet_be.exception.CustomException;
 import com.example.starlet_be.exception.ErrorCode;
