+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/email/dto/request/EmailAddressDto.java
@@ -1,4 +1,4 @@
-package com.example.starlet_be.domains.email.dto;
+package com.example.starlet_be.domains.email.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
