+++ Starlet_BE/src/main/java/com/example/starlet_be/domains/mypage/mapper/ConstellationMapper.java
@@ -1,9 +1,9 @@
 package com.example.starlet_be.domains.mypage.mapper;
 
 import com.example.starlet_be.domains.constellation.entity.Constellation;
-import com.example.starlet_be.domains.connection.dto.StarryNightConnectionDto;
+import com.example.starlet_be.domains.connection.dto.response.StarryNightConnectionDto;
 import com.example.starlet_be.domains.star.dto.response.StarryNightStarDto;
-import com.example.starlet_be.domains.constellation.dto.StarryNightConstellationDto;
+import com.example.starlet_be.domains.constellation.dto.response.StarryNightConstellationDto;
 import org.springframework.stereotype.Component;
 
 import java.util.Collections;
